move all shared-api single file folder into middlewares folder
