+++ apps/api/src/index.ts
@@ -3,9 +3,9 @@ import { cors } from "@elysiajs/cors";
 import { openapi } from "@elysiajs/openapi";
 import Elysia from "elysia";
 
-import { authService } from "@hebo/shared-api/auth/auth-service";
-import { corsConfig } from "@hebo/shared-api/cors/cors-config";
-import { prismaErrors } from "@hebo/shared-api/errors/prisma-errors";
+import { authService } from "@hebo/shared-api/middlewares/auth/auth-service";
+import { corsConfig } from "@hebo/shared-api/middlewares/cors-config";
+import { prismaErrors } from "@hebo/shared-api/middlewares/prisma-errors";
 
 import { agentsModule } from "./modules/agents";
 import { branchesModule } from "./modules/branches";
