+++ apps/api/src/modules/branches.ts
@@ -7,8 +7,8 @@ import {
   branchesInputUpdate,
 } from "@hebo/database/src/generated/prismabox/branches";
 import { createSlug } from "@hebo/database/src/utils/create-slug";
-import { authService } from "@hebo/shared-api/auth/auth-service";
-import { dbClient } from "@hebo/shared-api/db-client/db-client";
+import { authService } from "@hebo/shared-api/middlewares/auth/auth-service";
+import { dbClient } from "@hebo/shared-api/middlewares/db-client";
 import supportedModels from "@hebo/shared-data/json/supported-models";
 
 export const supportedModelsUnion = t.Union(
