+++ packages/shared-api/middlewares/auth/auth-service.ts
