+++ packages/shared-api/middlewares/auth/stack-auth.ts
